See Description
* Renamed WmaEncoder to WmaFileWriter and removed constructor overloads except file and stream.
* Renamed AcmEncoder to AcmFileWriter and removed constructor overloads except file and stream.
* RecordDevice implements IDisposable.
* WmaFileWriter.Handle is private.
* WaveFormat is used in the constructor of WmaFileWriter.
